requirements/base.txt
Django==2.2.15
Pillow==7.2.0
psycopg2==2.8.5
django-waffle==1.0
djangorestframework==3.11.0
django-import-export==2.3.0
django-environ==0.4.5
django-treebeard==4.3.1
django-tenant-schemas==1.10
Wand==0.6.2
XlsxWriter==1.3.1
django-excel-response==2.0.5
drf-flex-fields==0.8.5
django-braces==1.14.0


# reportlab==3.1.8
# requests==2.18.4